craete pixel processing method & separate pad util
